adding validation for minimum/maximum slots, moving helpers to separate file so validation method can access
